Add support for Samsung devices
Remove check for phonebook_sharing_message_confirm_pin as that isn't used on all devices.

Added option of checking the message view in addition to the title to see if the device if a go plus.

Changed logic to handle cases where either the title or message aren't used. OnePlus doesn't use message at all.